move snackbar error message to utils
